Alterações

+++ routes/web.php
@@ -25,18 +25,22 @@ Route::view('admin/home', 'home')->name('admin.home')->middleware('is_admin');
 Route::view('admin/dashboard', 'admin.HomeAdmin')->name('admin.dashboard')->middleware('is_admin');
 Route::view('admin/adicionarProduto', 'admin.adicionarProduto')->name('admin.adcProduto')->middleware('is_admin');
 Route::view('admin/adicionarUsuario', 'admin.Adicionarusuario')->name('admin.adcUsuario')->middleware('is_admin');
-Route::view('admin/listarUsuario', 'admin.listarUsuarios')->name('admin.listarUsuario')->middleware('is_admin');
 Route::view('admin/configuracoes/loja', 'admin.configLoja')->name('admin.configLoja')->middleware('is_admin');
 Route::view('admin/configuracoes/pagamento', 'admin.configPagamento')->name('admin.configPagamento')->middleware('is_admin');
 
 // Metodos de Product
-Route::post('admin/cadastroProduto', 'ProductController@cadastroDeProduto')->name('admin.adicionaProduto')->middleware('is_admin');
-Route::get('admin/alterarProduto', 'ProductController@alterarProduto')->name('admin.altProduto')->middleware('is_admin');
-Route::delete('admin/deleteProduto/{id}', 'ProductController@deleteProduto')->name('admin.deletarProduto')->middleware('is_admin');
-Route::get('admin/editarProduto/{id}', 'ProductController@editarProduto')->name('admin.EditarProduto')->middleware('is_admin');
-Route::put('admin/updateProduto/{id}', 'ProductController@updateProduto')->name('admin.updateProduto')->middleware('is_admin');
+Route::post('admin/cadastroProduto', 'Admin\ProductController@cadastroDeProduto')->name('admin.CadastroProduto')->middleware('is_admin');
+Route::get('admin/alterarProduto', 'Admin\ProductController@alterarProduto')->name('admin.altProduto')->middleware('is_admin');
+Route::delete('admin/deleteProduto/{id}', 'Admin\ProductController@deleteProduto')->name('admin.deletarProduto')->middleware('is_admin');
+Route::get('admin/editarProduto/{id}', 'Admin\ProductController@editarProduto')->name('admin.EditarProduto')->middleware('is_admin');
+Route::put('admin/updateProduto/{id}', 'Admin\ProductController@updateProduto')->name('admin.updateProduto')->middleware('is_admin');
 
 // Metodos de Usuarios
+
+Route::post('admin/cadastroUsuario', 'Admin\UserController@cadastroUsuario')->name('admin.AdicionarUsuario')->middleware('is_admin');
+Route::get('admin/listarUsuario', 'Admin\UserController@listarUsuarios')->name('admin.listarUsuario')->middleware('is_admin');
+Route::delete('admin/deleteUsuario/{id}','Admin\UserController@deletarUsuarios')->name('admin.deletarUsuario')->middleware('is_admin');
+
